+++ usr.sbin/vmd/vioblk.c
@@ -399,7 +399,7 @@ dev_dispatch_vm(int fd, short event, void *arg)
 	struct imsgev		*iev = &dev->async_iev;
 	struct imsgbuf		*ibuf = &iev->ibuf;
 	struct imsg	 	 imsg;
-	ssize_t			 n = 0;
+	int			 n = 0;
 	int			 verbose;
 	uint32_t		 type;
 
@@ -467,7 +467,7 @@ handle_sync_io(int fd, short event, void *arg)
 	struct imsgbuf *ibuf = &iev->ibuf;
 	struct viodev_msg msg;
 	struct imsg imsg;
-	ssize_t n;
+	int n;
 	int deassert = 0;
 
 	if (event & EV_READ) {
